Add distributed strategy, Add --gpus and --n_gpus parameter options

+++ utils/gpu.py
@@ -2,14 +2,18 @@ import os
 import numpy as np
 
 
-def find_free_gpus():
-    if os.system('nvidia-smi -q -d Memory |grep -A4 GPU|grep Free >tmp') == 0:
-        memory_available = [int(x.split()[2]) for x in open('tmp', 'r').readlines()]
-        gpus_most_idle = np.argsort(memory_available)[::-1]
-        os.remove('tmp')
-        return gpus_most_idle
+def get_cuda_visible_devices(selected_gpus, n_gpus):
+    gpus = []
+    if selected_gpus == "auto":
+        if os.system('nvidia-smi -q -d Memory |grep -A4 GPU|grep Free >tmp') == 0:
+            memory_available = [int(x.split()[2]) for x in open('tmp', 'r').readlines()]
+            gpus_most_idle = list(np.argsort(memory_available)[::-1])
+            os.remove('tmp')
+            if n_gpus != -1:
+                gpus_most_idle = gpus_most_idle[:n_gpus]
+            gpus = gpus_most_idle
+    else:
+        assert(n_gpus == "all", "Can only use one of --gpus and --n_gpus.")
+        gpus = selected_gpus
+    return ','.join(gpus) if gpus else ''
 
-
-def find_free_gpu():
-    gpus = find_free_gpus()
-    return str(gpus[0]) if gpus else ""
